fix: AlunoNotaService and call methods from controller
