+++ src/main/java/com/fmt/educonnect/interfaces/AlunoNotaInterface.java
@@ -1,5 +1,6 @@
 package com.fmt.educonnect.interfaces;
 
+import com.fmt.educonnect.controllers.dtos.requests.RequestAlunoDTO;
 import com.fmt.educonnect.datasource.entities.AlunoEntity;
 import com.fmt.educonnect.datasource.entities.NotaEntity;
 
@@ -11,4 +12,6 @@ public interface AlunoNotaInterface {
     List<NotaEntity> buscarNotasPorIdAluno(AlunoEntity alunoEntity);
 
     AlunoEntity buscarAlunoPorId(Long id);
+
+    AlunoEntity criarAluno(RequestAlunoDTO requestAlunoDTO);
 }
